preview changelog and confirm before publishing
