+++ cli/src/cli.ts
@@ -22,7 +22,6 @@ program
     .description('Generate changelog based on git commit history')
     .option('--from <ref>', 'Start ref (commit, tag, or branch)')
     .option('--to <ref>', 'End ref (commit, tag, or branch)', 'HEAD')
-    .option('--preview', 'Preview changelog in console')
     .option('--output <file>', 'Output file (default: stdout)')
     .option("--publish", 'Publish your changelog to changelog site')
     .action((opts) => {
